requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.1
aiosignal==1.2.0
apscheduler==3.9.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
fake-useragent==0.1.11
frozenlist==1.3.0
gunicorn==20.1.0
idna==3.3
multidict==6.0.2
mysql-connector==2.2.9
pymysql==1.0.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
setuptools==63.1.0
six==1.16.0
soupsieve==2.3.2.post1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
yarl==1.7.2
